+++ spring-data-elasticsearch-rest/src/main/java/com/github/ydespreaux/spring/data/elasticsearch/core/geo/MultiPolygonShape.java
@@ -44,7 +44,7 @@ import java.util.Objects;
 public class MultiPolygonShape extends AbstractShape implements CoordinatesShape<Coordinate[][][]> {
 
     @JsonIgnore
-    private transient List<PolygonShape> shapes;
+    private List<PolygonShape> shapes;
 
     private GeoShapeOrientation orientation;
 
@@ -78,7 +78,7 @@ public class MultiPolygonShape extends AbstractShape implements CoordinatesShape
     @Override
     public Coordinate[][][] getCoordinates() {
         if (shapes == null) {
-            return null;
+            return new Coordinate[0][][];
         }
         Coordinate[][][] coordinates = new Coordinate[this.shapes.size()][][];
         for (int i = 0; i < this.shapes.size(); i++) {